repo/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
SQLAlchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0
asyncpg==0.29.0
python-dotenv==1.0.0

PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
PyMuPDF==1.23.8

sentence-transformers==2.2.2
huggingface-hub==0.25.2
faiss-cpu==1.7.4

openai==1.3.7
requests==2.31.0
httpx==0.25.2

pytest==7.4.3
pytest-asyncio==0.21.1

reportlab==4.0.7
prometheus-client==0.19.0

websockets==12.0
sse-starlette==1.8.2
slowapi==0.1.9

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3

groq==0.4.2
tavily-python==0.3.0

psutil
